assets: scripts: created feather fall script
Create a script to control the feather fall. Currently, there is no way
to simulate the movements that a feather does when it's falling with air
resistance. Since the idea of the project is to compare the movement
both with and without air resistance, this is much needed.

It actually doesn't correctly simulates physically the forces envolved
in the movement, but actually with two mathematical equations: a sin
equation to simulate the movement in one way and another in the x axis
and a simple constant velocity equation to calculate the coordinate in
the y axis.